+++ packages/create-buntok/src/index.ts
@@ -165,7 +165,8 @@ app.get("/", (ctx) => {
 export default app;
 `;
 
-const ENV_TS_TEMPLATE = `import { App, z } from "@buntok/core";
+const ENV_TS_TEMPLATE = `import { App } from "@buntok/core";
+import { z } from "@buntok/core/middlewares/validator";
 
 export const env = App.validateEnv({
 	PORT: z.coerce.number().default(1212),